chore: move private snotra route
