+++ src/routes/index.ts
@@ -5,7 +5,6 @@ import alerts from './alerts';
 import redirector from './redirector';
 import devcards from './devcards';
 import privateRoutes from './private';
-import privateSnotraRoutes from './privateSnotra';
 import whoami from './whoami';
 import notifications from './notifications';
 import boot from './boot';
@@ -29,7 +28,6 @@ export default async function (fastify: FastifyInstance): Promise<void> {
   fastify.register(devcards, { prefix: '/devcards' });
   if (process.env.ENABLE_PRIVATE_ROUTES === 'true') {
     fastify.register(privateRoutes, { prefix: '/p' });
-    fastify.register(privateSnotraRoutes, { prefix: '/p/snotra' });
   }
   fastify.register(whoami, { prefix: '/whoami' });
   fastify.register(boot, { prefix: '/boot' });
