+++ src/routes/private.ts
@@ -14,6 +14,7 @@ import {
 } from '../entity/user/utils';
 import { queryReadReplica } from '../common/queryReadReplica';
 
+import snotra from './private/snotra';
 import rpc from './private/rpc';
 import { connectRpcPlugin } from '../common/connectRpc';
 
@@ -177,6 +178,7 @@ export default async function (fastify: FastifyInstance): Promise<void> {
     });
   });
 
+  fastify.register(snotra, { prefix: '/snotra' });
   fastify.register(connectRpcPlugin, {
     routes: rpc,
     prefix: '/rpc',
